requirements.txt
aerich==0.5.3
aiomysql==0.0.21
aiosqlite==0.16.1
click==8.0.1
ddlparse==1.9.0
dictdiffer==0.8.1
ecdsa==0.17.0
fastapi==0.65.2
iso8601==0.1.14
loguru==0.5.3
packaging==20.9
pyasn1==0.4.8
pydantic==1.8.2
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.17.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
